+++ app/src/main/java/com/fibelatti/photowidget/configure/SavePhotoWidgetUseCase.kt
@@ -111,12 +111,17 @@ class SavePhotoWidgetUseCase @Inject constructor(
 
         photoWidgetStorage.saveWidgetOpacity(
             appWidgetId = appWidgetId,
-            opacity = photoWidget.opacity,
+            opacity = photoWidget.colors.opacity,
         )
 
-        photoWidgetStorage.saveWidgetBlackAndWhite(
+        photoWidgetStorage.saveWidgetSaturation(
             appWidgetId = appWidgetId,
-            value = photoWidget.blackAndWhite,
+            saturation = photoWidget.colors.saturation,
+        )
+
+        photoWidgetStorage.saveWidgetBrightness(
+            appWidgetId = appWidgetId,
+            brightness = photoWidget.colors.brightness,
         )
 
         photoWidgetStorage.saveWidgetOffset(